backtrack-agent/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
aiohttp==3.9.5
docker==7.1.0
kubernetes==30.1.0
statsmodels==0.14.2
scikit-learn==1.5.0
numpy==1.26.4
prometheus-client==0.20.0
python-dotenv==1.0.1
